# switchbook-app/public/openapi/partner-v1.yaml
openapi: 3.1.0
info:
  title: SwitchBook Partner API
  version: 1.0.0
  description: Canonical approved mechanical-switch catalog and moderated contribution API.
  license: { name: Proprietary integration terms, identifier: LicenseRef-SwitchBook-Partner }
servers:
  - url: https://switchbook.app/api/v1
  - url: https://sandbox.switchbook.app/api/v1
security:
  - applicationKey: []
paths:
...
  /submissions:
    get:
      summary: List the current user's partner submissions
      operationId: listSubmissions
      security: [ { oauth: [submissions:read] } ]
      responses: { '200': { description: Current user's submissions }, '401': { $ref: '#/components/responses/Error' } }
    post:
      summary: Propose a missing switch for moderation
      operationId: createSubmission
      security: [ { oauth: [submissions:write] } ]
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SwitchProposal' } } } }
      responses: { '202': { description: Proposal accepted for moderation }, '409': { description: Possible duplicate or idempotency conflict } }
...
  /catalog/switches/{id}/corrections:
    post:
      summary: Suggest a correction to an active canonical record
      operationId: createCorrection
      security: [ { oauth: [corrections:write] } ]
      parameters: [ { $ref: '#/components/parameters/SwitchId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CorrectionProposal' } } } }
      responses: { '202': { description: Correction accepted for moderation }, '400': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    applicationKey: { type: apiKey, in: header, name: X-API-Key }
    oauth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://switchbook.app/oauth2/auth
          tokenUrl: https://switchbook.app/oauth2/token
          refreshUrl: https://switchbook.app/oauth2/token
          scopes:
            openid: Authenticate the SwitchBook account
            offline_access: Issue a rotating refresh token
            profile:read: Read linked basic profile
            submissions:write: Create catalog proposals
            submissions:read: Read own moderation status
            corrections:write: Suggest canonical corrections
  parameters:
    SwitchId: { name: id, in: path, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 200 } }
  responses:
    Error: { description: Structured error, content: { application/json: { schema: { type: object, properties: { error: { $ref: '#/components/schemas/Error' } } } } } }
  schemas:
    Error: { type: object, required: [code, message, requestId], properties: { code: { type: string }, message: { type: string }, requestId: { type: string }, details: {} } }
...
    PhotoProposal: { type: object, required: [url, alt], properties: { url: { type: string, format: uri }, alt: { type: string, minLength: 1, maxLength: 500 }, sourceUrl: { type: string, format: uri }, license: { type: string }, attribution: { type: string } } }
    SwitchProposal:
      type: object
      required: [name, manufacturer, submissionNotes]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        manufacturer: { type: string, minLength: 1, maxLength: 150 }
        chineseName: { type: [string, 'null'] }
        type: { type: [string, 'null'], enum: [LINEAR, TACTILE, CLICKY, SILENT_LINEAR, SILENT_TACTILE, MOUSE, null] }
        technology: { type: [string, 'null'], enum: [MECHANICAL, OPTICAL, MAGNETIC, INDUCTIVE, ELECTRO_CAPACITIVE, null] }
        initialForce: { type: [number, 'null'] }
        actuationForce: { type: [number, 'null'] }
        tactileForce: { type: [number, 'null'] }
        bottomOutForce: { type: [number, 'null'] }
        preTravel: { type: [number, 'null'] }
        bottomOut: { type: [number, 'null'] }
        springWeight: { type: [string, 'null'] }
        springLength: { type: [string, 'null'] }
        compatibility: { type: [string, 'null'] }
        photos: { type: array, maxItems: 8, items: { $ref: '#/components/schemas/PhotoProposal' } }
        submissionNotes: { type: string, minLength: 10, maxLength: 4000 }
        confirmNotDuplicate: { type: boolean, default: false }
    CorrectionProposal: { type: object, required: [changes, reason], properties: { changes: { type: object, minProperties: 1, additionalProperties: true }, reason: { type: string, minLength: 10, maxLength: 4000 } } }